requirements_multi_gpu.txt
# Requirements para BGE-M3 Multi-GPU Optimizado (4x L40S)
# Versiones específicas para máximo rendimiento

# Core ML/DL
torch>=2.6.0  # Requerido para evitar vulnerabilidad CVE-2025-32434
transformers>=4.35.0
FlagEmbedding>=1.2.10
sentence-transformers>=2.2.2
accelerate>=0.25.0

# CUDA y optimizaciones
nvidia-ml-py3>=7.352.0
# flash-attn>=2.3.0  # Opcional, requiere compilación en Windows
# triton>=2.1.0  # Opcional, no disponible en Windows

# Base de datos y vectores
qdrant-client>=1.6.4
# pymongo>=4.6.0  # Ya no necesario - usando JSON
# motor>=3.3.2  # Ya no necesario - usando JSON

# Async y concurrencia
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
uvloop>=0.19.0  # Faster event loop (Linux/Mac)

# Monitoring y utilidades
psutil>=5.9.6
GPUtil>=1.4.0
tqdm>=4.66.1
rich>=13.7.0  # Beautiful terminal output
wandb>=0.16.0  # Optional: experiment tracking

# Data processing
numpy>=1.24.0
pandas>=2.1.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
omegaconf>=2.3.0

# Logging y debugging
loguru>=0.7.2
memory-profiler>=0.61.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development (optional)
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
